+++ src/lmprof_stack.h
@@ -7,7 +7,7 @@ typedef struct lmprof_stack lmprof_stack;
 
 lmprof_stack* lmprof_stack_create  ();
 void          lmprof_stack_destroy (lmprof_stack* s);
-int           lmprof_stack_push    (lmprof_stack *s, int element);
-int           lmprof_stack_pop     (lmprof_stack *s);
+int           lmprof_stack_push    (lmprof_stack *s, size_t element);
+size_t        lmprof_stack_pop     (lmprof_stack *s);
  
 #endif
